mcp/rag_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0